test(socket-factory): move flow-socket test to integration test
Exercise PerfUdpSocket through its public API from tests/ rather than an
in-module test. The round-trip runs on every platform; the flow-socket
assertion stays Apple-gated and reads the count via a doc(hidden) accessor.
